Modified Entity, Controllers, and Services

+++ backend/src/main/java/edu/cit/projectsync/Repository/UserRepository.java
@@ -1,13 +1,11 @@
 package edu.cit.projectsync.Repository;
 
-import java.util.Optional;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import edu.cit.projectsync.Entity.UserEntity;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
     UserEntity findByEmail(String email);
 }
